+++ src/AbstractValue.php
@@ -1,4 +1,5 @@
 <?php
+
 /*
  * Copyright 2022 Cloud Creativity Limited
  *