+++ src/ValueTrait.php
@@ -23,11 +23,8 @@ trait ValueTrait
     /**
      * @var TValue
      */
-    protected mixed $value;
+    public readonly mixed $value;
 
-    /**
-     * @return string
-     */
     public function __toString(): string
     {
         return $this->toString();
@@ -35,8 +32,6 @@ trait ValueTrait
 
     /**
      * Fluent to string method.
-     *
-     * @return string
      */
     public function toString(): string
     {
@@ -58,8 +53,7 @@ trait ValueTrait
     /**
      * Is the value any of the provided values?
      *
-     * @param mixed ...$values
-     * @return bool
+     * @param  mixed  ...$values
      */
     public function is(...$values): bool
     {
@@ -76,20 +70,14 @@ trait ValueTrait
         return false;
     }
 
-    /**
-     * @return bool
-     */
     public function isEmpty(): bool
     {
         return empty($this->value);
     }
 
-    /**
-     * @return bool
-     */
     public function isNotEmpty(): bool
     {
-        return !$this->isEmpty();
+        return ! $this->isEmpty();
     }
 
     /**
@@ -102,9 +90,6 @@ trait ValueTrait
 
     /**
      * Does the value match the provided value?
-     *
-     * @param mixed $value
-     * @return bool
      */
     protected function matches(mixed $value): bool
     {
@@ -121,8 +106,6 @@ trait ValueTrait
 
     /**
      * Should strict comparison be used for comparing values?
-     *
-     * @return bool
      */
     protected function useStrict(): bool
     {