+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. This projec
 ### Changed
 
 - Minimum PHP version is now 8.2.
+- Added generics to the value interface, abstract class and trait. This will be breaking if you are running PHPStan.
 
 ## [2.2.0] - 2022-03-04